Factor out creator behaviour from View.
